requirements.txt
PyQt5
simpleaudio
edge-tts
google-generativeai
SpeechRecognition
pyaudio
pygame
pillow
keyboard
